updated tests and stuff
